Remove group check on installation creation (#43)
This removes the API call to the provisioner that was originally
used to ensure the group ID was valid. The code worked as intended,
but only served the purpose of providing a cleaner error message
if the group did not exist. A similar error will now be returned
later when the installation creation is attempted.

This change is being made to reduce code complexity and to remove
the requirement for the plugin to hit the `group` API endpoints at
all. This allows for stricter authentication scope if desired.

Also, the plugin now properly handles installation webhooks where
the old state was `creation-final-tasks`.