Change `upgrade` command to `update`
The `upgrade` plugin command contains a lot of functionality that
doesn't result in a meaningful server upgrade. This change of the
command name to `update` better reflects what it does while also
bringing it closer in line with the provisioner naming. A helper
`upgrade` function exists to direct users to the new command.