Display Service Environment configuration in RHS
This change is the first step in better handling of service
environment configuration in the plugin. The installation environment
variables are now examined to expose which service environment
value the test server will use. This is displayed in the plugin RHS.
The creation date of the server is now also displayed for
informational purposes.