Add scheduled deletion for installations (#170)
The plugin now supports optional installation deletion scheduling.
The deletion time is set via plugin configuration in the form of
the number of hours installations should exist for. If this is set
the webapp RHS will show the value to users for those installations.
If an installation is deletion locked then this is hidden to
reduce confusion.

This change also includes the new `pluginctl` tooling for an
improved dev experience.